+++ api/tests/identity-access-management/unit/infrastructure/serializers/jsonapi/user-login-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/user-login-serializer.js';
-import { UserLogin } from '../../../../../src/identity-access-management/domain/models/UserLogin.js';
-import { expect } from '../../../../test-helper.js';
+import { UserLogin } from '../../../../../../src/identity-access-management/domain/models/UserLogin.js';
+import * as serializer from '../../../../../../src/identity-access-management/infrastructure/serializers/jsonapi/user-login-serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | user-login-serializer', function () {
   describe('#serialize', function () {
